refactor: config 파일 ai_report_data로 이름변경
